requirements.txt
slack_bolt
google-genai
python-dotenv